requirements.txt
Flask
gunicorn
tensorflow-cpu
numpy
opencv-python-headless
Pillow
